Inital version of thunks generation script for EGL/GLES proxy libs
We can provide small proxy shared libraries for EGL/GLESv1/GLESv2
and decide at runtime which variants of the blob drivers need to
be loaded to do the real job.

Yes, this smells kinda like libhybris, but there seems to be no
better option to take care of all the quirks in the r3p0 and r3p2
mali400 blobs. And more importantly, allow to switch between
r3p0 and r3p2 mali drivers by just replacing the kernel (no
changes in the userspace required).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>